requirements.txt
transformers==4.41.2
torch
sentencepiece
python-dotenv
